Moved the adapter to a separate file
